more debugging

+++ scripts/install_awscli.sh
@@ -1,5 +1,8 @@
 #!/bin/bash -uex
-rm ~/.curlrc
+
+if [ -f ~/.curlrc ]; then
+    rm ~/.curlrc
+fi
 
 rm 
 _arch="$(uname -m)"
